fix: fixed bug that only updated checkbox value on screen rotation
Now preference checkbox value is updated corretly. Code that updated the value was moved to onResume method. Since onResume is called after onCreate and after SettingsActivity is closed it works correctly.
As a side note: in order to make onOptionsItemSelected work in fragment, method has to be overriden and also under the items that are overrided you need to return false in the MainActivity.